+++ src/components/ListingCard/ListingCard.js
@@ -1,11 +1,11 @@
 import React, { Component } from 'react';
-import { string, func } from 'prop-types';
+import { array, string, func } from 'prop-types';
 import { FormattedMessage, intlShape, injectIntl } from '../../util/reactIntl';
 import classNames from 'classnames';
 import { lazyLoadWithDimensions } from '../../util/contextHelpers';
 import { LINE_ITEM_DAY, LINE_ITEM_NIGHT, propTypes } from '../../util/types';
 import { formatMoney } from '../../util/currency';
-import { ensureListing, ensureUser } from '../../util/data';
+import { ensureListing } from '../../util/data';
 import { richText } from '../../util/richText';
 import { createSlug } from '../../util/urlHelpers';
 import config from '../../config';
@@ -34,6 +34,10 @@ const priceData = (price, intl) => {
   return {};
 };
 
+const getCertificateInfo = (certificateConfig, key) => {
+  return certificateConfig.find(c => c.key === key);
+};
+
 class ListingImage extends Component {
   render() {
     return <ResponsiveImage {...this.props} />;
@@ -42,17 +46,26 @@ class ListingImage extends Component {
 const LazyImage = lazyLoadWithDimensions(ListingImage, { loadAfterInitialRendering: 3000 });
 
 export const ListingCardComponent = props => {
-  const { className, rootClassName, intl, listing, renderSizes, setActiveListing } = props;
+  const {
+    className,
+    rootClassName,
+    intl,
+    listing,
+    renderSizes,
+    certificateConfig,
+    setActiveListing,
+  } = props;
   const classes = classNames(rootClassName || css.root, className);
   const currentListing = ensureListing(listing);
   const id = currentListing.id.uuid;
-  const { title = '', price } = currentListing.attributes;
+  const { title = '', price, publicData } = currentListing.attributes;
   const slug = createSlug(title);
-  const author = ensureUser(listing.author);
-  const authorName = author.attributes.profile.displayName;
   const firstImage =
     currentListing.images && currentListing.images.length > 0 ? currentListing.images[0] : null;
 
+  const certificate = publicData
+    ? getCertificateInfo(certificateConfig, publicData.certificate)
+    : null;
   const { formattedPrice, priceTitle } = priceData(price, intl);
 
   const unitType = config.bookingUnitType;
@@ -98,8 +111,10 @@ export const ListingCardComponent = props => {
               longWordClass: css.longWord,
             })}
           </div>
-          <div className={css.authorInfo}>
-            <FormattedMessage id="ListingCard.hostedBy" values={{ authorName }} />
+          <div className={css.certificateInfo}>
+            {certificate && !certificate.hideFromListingInfo ? (
+              <span>{certificate.label}</span>
+            ) : null}
           </div>
         </div>
       </div>
@@ -111,12 +126,14 @@ ListingCardComponent.defaultProps = {
   className: null,
   rootClassName: null,
   renderSizes: null,
+  certificateConfig: config.custom.certificate,
   setActiveListing: () => null,
 };
 
 ListingCardComponent.propTypes = {
   className: string,
   rootClassName: string,
+  certificateConfig: array,
   intl: intlShape.isRequired,
   listing: propTypes.listing.isRequired,