+++ src/components/SelectMultipleFilter/SelectMultipleFilter.example.js
@@ -3,41 +3,15 @@ import { withRouter } from 'react-router-dom';
 import SelectMultipleFilter from './SelectMultipleFilter';
 import { stringify, parse } from '../../util/urlHelpers';
 
-const URL_PARAM = 'pub_amenities';
+const URL_PARAM = 'pub_yogaStyles';
 
 const options = [
-  {
-    key: 'towels',
-    label: 'Towels',
-  },
-  {
-    key: 'bathroom',
-    label: 'Bathroom',
-  },
-  {
-    key: 'swimming_pool',
-    label: 'Swimming pool',
-  },
-  {
-    key: 'own_drinks',
-    label: 'Own drinks allowed',
-  },
-  {
-    key: 'jacuzzi',
-    label: 'Jacuzzi',
-  },
-  {
-    key: 'audiovisual_entertainment',
-    label: 'Audiovisual entertainment',
-  },
-  {
-    key: 'barbeque',
-    label: 'Barbeque',
-  },
-  {
-    key: 'own_food_allowed',
-    label: 'Own food allowed',
-  },
+  { key: 'ashtanga', label: 'Ashtanga' },
+  { key: 'hatha', label: 'Hatha' },
+  { key: 'kundalini', label: 'Kundalini' },
+  { key: 'restorative', label: 'Restorative' },
+  { key: 'vinyasa', label: 'Vinyasa' },
+  { key: 'yin', label: 'yin' },
 ];
 
 const handleSubmit = (urlParam, values, history) => {
@@ -46,19 +20,19 @@ const handleSubmit = (urlParam, values, history) => {
   history.push(`${window.location.pathname}${queryParams}`);
 };
 
-const AmenitiesFilterPopup = withRouter(props => {
+const YogaStylesFilterPopup = withRouter(props => {
   const { history, location } = props;
 
   const params = parse(location.search);
-  const amenities = params[URL_PARAM];
-  const initialValues = !!amenities ? amenities.split(',') : [];
+  const yogaStyles = params[URL_PARAM];
+  const initialValues = !!yogaStyles ? yogaStyles.split(',') : [];
 
   return (
     <SelectMultipleFilter
       id="SelectMultipleFilterPopupExample"
-      name="amenities"
+      name="yogaStyles"
       urlParam={URL_PARAM}
-      label="Amenities"
+      label="yogaStyles"
       onSubmit={(urlParam, values) => handleSubmit(urlParam, values, history)}
       showAsPopup={true}
       liveEdit={false}
@@ -69,25 +43,25 @@ const AmenitiesFilterPopup = withRouter(props => {
   );
 });
 
-export const AmenitiesFilterPopupExample = {
-  component: AmenitiesFilterPopup,
+export const YogaStylesFilterPopupExample = {
+  component: YogaStylesFilterPopup,
   props: {},
   group: 'filters',
 };
 
-const AmenitiesFilterPlain = withRouter(props => {
+const YogaStylesFilterPlain = withRouter(props => {
   const { history, location } = props;
 
   const params = parse(location.search);
-  const amenities = params[URL_PARAM];
-  const initialValues = !!amenities ? amenities.split(',') : [];
+  const yogaStyles = params[URL_PARAM];
+  const initialValues = !!yogaStyles ? yogaStyles.split(',') : [];
 
   return (
     <SelectMultipleFilter
       id="SelectMultipleFilterPlainExample"
-      name="amenities"
+      name="yogaStyles"
       urlParam={URL_PARAM}
-      label="Amenities"
+      label="yogaStyles"
       onSubmit={(urlParam, values) => {
         handleSubmit(urlParam, values, history);
       }}
@@ -99,8 +73,8 @@ const AmenitiesFilterPlain = withRouter(props => {
   );
 });
 
-export const AmenitiesFilterPlainExample = {
-  component: AmenitiesFilterPlain,
+export const YogaStylesFilterPlainExample = {
+  component: YogaStylesFilterPlain,
   props: {},
   group: 'filters',
 };