+++ src/containers/ListingPage/SectionHeading.js
@@ -6,13 +6,17 @@ import config from '../../config';
 
 import css from './ListingPage.css';
 
+const getCertificateInfo = (certificateConfig, key) => {
+  return certificateConfig.find(c => c.key === key);
+};
+
 const SectionHeading = props => {
   const {
     priceTitle,
     formattedPrice,
     richTitle,
-    category,
-    hostLink,
+    listingCertificate,
+    certificateConfig,
     showContactUser,
     onContactUser,
   } = props;
@@ -27,6 +31,8 @@ const SectionHeading = props => {
     ? 'ListingPage.perDay'
     : 'ListingPage.perUnit';
 
+  const certificate = getCertificateInfo(certificateConfig, listingCertificate);
+  const showCertificate = certificate && !certificate.hideFromListingInfo;
   return (
     <div className={css.sectionHeading}>
       <div className={css.desktopPriceContainer}>
@@ -40,11 +46,10 @@ const SectionHeading = props => {
       <div className={css.heading}>
         <h1 className={css.title}>{richTitle}</h1>
         <div className={css.author}>
-          {category}
-          <FormattedMessage id="ListingPage.hostedBy" values={{ name: hostLink }} />
+          {showCertificate ? <span>{certificate.label}</span> : null}
           {showContactUser ? (
             <span className={css.contactWrapper}>
-              <span className={css.separator}>•</span>
+              {showCertificate ? <span className={css.separator}>•</span> : null}
               <InlineTextButton rootClassName={css.contactLink} onClick={onContactUser}>
                 <FormattedMessage id="ListingPage.contactUser" />
               </InlineTextButton>