+++ src/marketplace-custom-config.js
@@ -2,46 +2,19 @@
  * Marketplace specific configuration.
  */
 
-export const amenities = [
-  {
-    key: 'towels',
-    label: 'Towels',
-  },
-  {
-    key: 'bathroom',
-    label: 'Bathroom',
-  },
-  {
-    key: 'swimming_pool',
-    label: 'Swimming pool',
-  },
-  {
-    key: 'own_drinks',
-    label: 'Own drinks allowed',
-  },
-  {
-    key: 'jacuzzi',
-    label: 'Jacuzzi',
-  },
-  {
-    key: 'audiovisual_entertainment',
-    label: 'Audiovisual entertainment',
-  },
-  {
-    key: 'barbeque',
-    label: 'Barbeque',
-  },
-  {
-    key: 'own_food_allowed',
-    label: 'Own food allowed',
-  },
+export const yogaStyles = [
+  { key: 'ashtanga', label: 'Ashtanga' },
+  { key: 'hatha', label: 'Hatha' },
+  { key: 'kundalini', label: 'Kundalini' },
+  { key: 'restorative', label: 'Restorative' },
+  { key: 'vinyasa', label: 'Vinyasa' },
+  { key: 'yin', label: 'Yin' },
 ];
 
-export const categories = [
-  { key: 'smoke', label: 'Smoke' },
-  { key: 'electric', label: 'Electric' },
-  { key: 'wood', label: 'Wood' },
-  { key: 'other', label: 'Other' },
+export const certificate = [
+  { key: 'none', label: 'None', hideFromFilters: true, hideFromListingInfo: true },
+  { key: '200h', label: 'Registered yoga teacher 200h' },
+  { key: '500h', label: 'Registered yoga teacher 500h' },
 ];
 
 // Price filter configuration